[update] Aboutページのレイアウト
