+++ app/javascript/stylesheets/application.scss
@@ -8,3 +8,4 @@
 @use "./public/registration-form";
 @use "./public/main-visual";
 @use "./public/sidebar";
+@use "./public/about";
