+++ app/javascript/stylesheets/global/_layout.scss
@@ -1,4 +1,5 @@
 @use "./variables" as *;
+@import url("https://fonts.googleapis.com/css2?family=Shippori+Mincho+B1:wght@800&display=swap");
 
 body {
   background-image: url($body-wallpaper);
