+++ app/views/public/homes/about.html.erb
@@ -1,11 +1,16 @@
-<section class="text-center">
-  <h1>About</h1>
-  <p>
-    ながのCAKEは全国発送可能なケーキ屋さんです。<br>
-    お子様からお年寄りの方までお召し上がりいただけるよう、<br>
-    様々なバリエーションをご用意しております。<br>
-    また、全国へ発送可能ですので「いつ」でも「どこ」でもご注文いただけます。<br>
-    ぜひご注文ください。<br>
-  </p>
-  <div></div>
-</section>
+<%= render "public/layouts/frame", type: :standard do %>
+  <div class="text-center my-5">
+    <h1 class="display-3 my-2">About</h1>
+    <p class="lead mb-5">
+      ながのCAKEは全国発送可能なケーキ屋さんです。<br>
+      お子様からお年寄りの方までお召し上がりいただけるよう、<br>
+      様々なバリエーションをご用意しております。<br>
+      また、全国へ発送可能ですので「いつ」でも「どこ」でもご注文いただけます。<br>
+      ぜひご注文ください。<br>
+    </p>
+    <div id="about-image-container">
+      <%= image_tag "", id: "about-image", draggable: false, width: 640, alt: "イメージ写真", decoding: "async", loading: "lazy" %>
+      <%= image_tag "logo.svg", id:"about-logo", draggable: false %>
+    </div>
+  </div>
+<% end %>
